update:新增了一些判断，减少程序错误的可能性

+++ src/main/java/com/java/demomp/plan/service/impl/PlanServiceImpl.java
@@ -10,6 +10,7 @@ import com.java.demomp.plan.service.DictParentService;
 import com.java.demomp.plan.service.DictService;
 import com.java.demomp.plan.service.PlanService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.java.demomp.util.PlanUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -164,7 +165,13 @@ public class PlanServiceImpl extends ServiceImpl<PlanMapper, Plan> implements Pl
         // 首先获得父id
         Integer parentId = baseMapper.selectById(id).getParentId();
 
-        // 二话不说，先删了再说
+        // 二话不说，先删了再说(刀下留人，如果删了父亲节点，儿子没归宿了)
+        // 所以要先获得孩子节点，如果没有孩子节点才能删除
+        List<Plan> children = baseMapper.selectList(new QueryWrapper<Plan>().eq("parent_id", id));
+        if(children == null || children.size() == 0){
+            return 0;
+        }
+        // 确保没有孩子才能删除
         int i = baseMapper.deleteById(id);
         while (parentId != null) {
             parentId = updateFatherByParentId(parentId);
@@ -216,6 +223,7 @@ public class PlanServiceImpl extends ServiceImpl<PlanMapper, Plan> implements Pl
                 // 下面2个是非必要属性
                 map.put("color",plan.getColor());
                 map.put("percent",plan.getPercent());
+                map.put("rank", PlanUtil.getRankStrByRank(plan.getRank()));
                 // 如果有儿子
                 List<Plan> children = baseMapper.selectList(new QueryWrapper<Plan>().eq("parent_id", plan.getId()));
                 if(children!= null && children.size()>0){
@@ -243,86 +251,6 @@ public class PlanServiceImpl extends ServiceImpl<PlanMapper, Plan> implements Pl
 //        return resultList;
 //    }
 
-
-    public void test(){
-
-        /**
-         *
-         *                  1                           8
-         *              2          3                9     10
-         *            4  5        6 7
-         *
-         *
-         *
-         */
-
-
-        // ----------------------  2添加  4和5
-        Map<String,Object> map2 = new HashMap<>();
-
-        Map<String,Object> map4 = new HashMap<>();
-        map4.put("id",4);
-
-        Map<String,Object> map5 = new HashMap<>();
-        map5.put("id",5);
-
-        map2.put("id",2);
-        List<Object> list2 = new ArrayList<>();
-        list2.add(map4);
-        list2.add(map5);
-        map2.put("children",list2);
-
-
-        //-------------------- 3添加  6和7
-
-        Map<String,Object> map3 = new HashMap<>();
-
-        Map<String,Object> map6 = new HashMap<>();
-        map6.put("id",6);
-
-        Map<String,Object> map7 = new HashMap<>();
-        map7.put("id",7);
-
-        map3.put("id",3);
-        List<Object> list3 = new ArrayList<>();
-        list3.add(map6);
-        list3.add(map7);
-        map3.put("children",list3);
-
-        //-------------------  1添加  2和3
-
-        Map<String,Object> map1 = new HashMap<>();
-        map1.put("id",1);
-        List<Object> list1 = new ArrayList<>();
-        list1.add(map2);
-        list1.add(map3);
-        map1.put("child",list1);
-
-
-        // -------------------8添加  9 和10
-        Map<String,Object> map8 = new HashMap<>();
-
-        Map<String,Object> map9 = new HashMap<>();
-        map9.put("id",9);
-
-        Map<String,Object> map10 = new HashMap<>();
-        map10.put("id",10);
-
-        map8.put("id",8);
-        List<Object> list8 = new ArrayList<>();
-        list8.add(map9);
-        list8.add(map10);
-        map8.put("children",list8);
-
-
-        // --------------------------------总体节点   添加1和8
-
-        List<Object> list = new ArrayList<>();
-        list.add(map1);
-        list.add(map8);
-        System.out.println(list);
-
-    }
 
 
     /**
